refactor(frontend): decompose PieceDetail, WorkflowState, PiecePhotoGallery (#406)
Extract 7 focused modules from the three largest frontend components to bring
each file under ~350 lines and isolate autosave, upload-queue, and caption logic:

- SectionCard.tsx — frosted-glass card wrapper (was inline in PieceDetail)
- EditableToggle.tsx — seal/unseal history toggle
- PieceNameEditor.tsx — inline-editable name heading with save-status integration
- ShowcaseStoryEditor.tsx — showcase story textarea with autosave
- ImageUploader.tsx — Cloudinary upload trigger (FAB + desktop button)
- useImageSaveQueue.ts — hook owning the Promise-chained image-save queue
- LightboxFooter.tsx — caption edit, move menu, thumbnail, and crop controls

Each new component ships with JSDoc and a Storybook story file.
BUILD.bazel updated: new files added to existing *_src js_library targets and
expected_coverage globs in vitest_test targets.

No behavior changes — all three parent component tests pass unchanged.
