+++ web/src/components/WorkflowState.tsx
@@ -3,20 +3,12 @@ import {
   useEffect,
   useMemo,
   useReducer,
-  useRef,
-  useState,
 } from "react";
-import PhotoCameraOutlinedIcon from "@mui/icons-material/PhotoCameraOutlined";
 import { useTheme } from "@mui/material";
 import {
   Box,
-  Button,
-  CircularProgress,
-  Fab,
   MenuItem,
-  Portal,
   TextField,
-  Typography,
   useMediaQuery,
 } from "@mui/material";
 import type { PieceDetail, PieceState } from "../util/types";
@@ -26,7 +18,6 @@ import {
   type UISchema,
   type UpdateStatePayload,
 } from "../util/api";
-import { openCloudinaryUploadWidget } from "../util/cloudinaryUpload";
 import {
   getCustomFieldDefinitions,
   getDefinitionsFromSchema,
@@ -38,11 +29,12 @@ import { useAutosave } from "./useAutosave";
 import { usePieceDetailSaveStatus } from "./usePieceDetailSaveStatus";
 import { useAsync } from "../util/useAsync";
 import {
-  type ImageEntry,
   buildDraftState,
   draftReducer,
   normalizeCustomFieldPayload,
 } from "./workflowStateDraft";
+import ImageUploader from "./ImageUploader";
+import { useImageSaveQueue } from "./useImageSaveQueue";
 
 type WorkflowStateProps = {
   initialPieceState: PieceState;
@@ -61,116 +53,6 @@ type WorkflowStateProps = {
   onChange?: (payload: UpdateStatePayload) => void;
 };
 
-// ── ImageUploader ─────────────────────────────────────────────────────────────
-
-type ImageUploaderProps = {
-  saving: boolean;
-  widgetLoading: boolean;
-  uploadError: string | null;
-  imageError: string | null;
-  mobile: boolean;
-  hidden?: boolean;
-  onUploadClick: () => void;
-};
-
-function ImageUploader({
-  saving,
-  widgetLoading,
-  uploadError,
-  imageError,
-  mobile,
-  hidden = false,
-  onUploadClick,
-}: ImageUploaderProps) {
-  const buttonDisabled = saving || widgetLoading;
-  const statusMessage = saving ? "Saving…" : "Upload Image";
-
-  return (
-    <Box sx={hidden ? { display: "none" } : undefined}>
-      {mobile ? (
-        <Portal>
-          <Fab
-            color="primary"
-            aria-label="Upload Image"
-            onClick={onUploadClick}
-            disabled={buttonDisabled}
-            hidden={hidden}
-            sx={{
-              display: hidden ? "none" : undefined,
-              position: "fixed",
-              right: 24,
-              bottom: 24,
-              zIndex: (theme) => theme.zIndex.speedDial,
-              boxShadow: (theme) => theme.shadows[8],
-            }}
-          >
-            {widgetLoading ? (
-              <CircularProgress aria-hidden size={20} color="inherit" />
-            ) : (
-              <PhotoCameraOutlinedIcon />
-            )}
-            <Box
-              component="span"
-              sx={{
-                position: "absolute",
-                width: 1,
-                height: 1,
-                p: 0,
-                m: -1,
-                overflow: "hidden",
-                clip: "rect(0 0 0 0)",
-                whiteSpace: "nowrap",
-                border: 0,
-              }}
-            >
-              {statusMessage}
-            </Box>
-          </Fab>
-        </Portal>
-      ) : (
-        <Portal
-          container={
-            (typeof document !== "undefined" &&
-              document.getElementById("piece-upload-trigger")) ||
-            null
-          }
-        >
-          <Button
-            variant="outlined"
-            size="small"
-            onClick={onUploadClick}
-            disabled={buttonDisabled}
-            hidden={hidden}
-            startIcon={
-              saving ? (
-                <CircularProgress size={14} color="inherit" />
-              ) : undefined
-            }
-            sx={{ display: hidden ? "none" : undefined, position: "relative" }}
-          >
-            <Box sx={{ opacity: widgetLoading ? 0 : 1 }}>{statusMessage}</Box>
-            {widgetLoading && (
-              <CircularProgress
-                aria-hidden
-                size={14}
-                color="inherit"
-                sx={{ position: "absolute" }}
-              />
-            )}
-          </Button>
-        </Portal>
-      )}
-      {(uploadError || imageError) && (
-        <Typography color="error" variant="body2" sx={{ mt: 1 }}>
-          {uploadError ?? imageError}
-        </Typography>
-      )}
-    </Box>
-  );
-}
-
-// ── WorkflowState ─────────────────────────────────────────────────────────────
-
 export default function WorkflowState({
   initialPieceState,
   pieceId,
@@ -185,8 +67,6 @@ export default function WorkflowState({
   disableAutosave = false,
   onChange,
 }: WorkflowStateProps) {
-  const [uploadError, setUploadError] = useState<string | null>(null);
-  const [widgetLoading, setWidgetLoading] = useState(false);
   const [draft, dispatch] = useReducer(
     draftReducer,
     initialPieceState,
@@ -236,11 +116,6 @@ export default function WorkflowState({
   const theme = useTheme();
   const isMobileLayout = useMediaQuery(theme.breakpoints.down("sm"));
 
-  const [savingImage, setSavingImage] = useState(false);
-  const [imageError, setImageError] = useState<string | null>(null);
-  const latestImagesRef = useRef<ImageEntry[]>(images);
-  const imageSaveQueueRef = useRef<Promise<void>>(Promise.resolve());
-
   const isDirty = notes !== baseState.notes || customFieldsDirty;
 
   const currentPayload = useMemo(
@@ -251,10 +126,6 @@ export default function WorkflowState({
     }),
     [notes, images, normalizedCustomFields],
   );
-
-  useEffect(() => {
-    latestImagesRef.current = images;
-  }, [images]);
 
   useEffect(() => {
     onDirtyChange?.(readOnly ? false : isDirty);
@@ -295,50 +166,23 @@ export default function WorkflowState({
   });
   const pieceDetailSaveStatus = usePieceDetailSaveStatus();
 
-  const saveUploadedImage = useCallback(
-    (newImage: ImageEntry) => {
-      setSavingImage(true);
-      setImageError(null);
-      const queuedSave = imageSaveQueueRef.current
-        .catch(() => undefined)
-        .then(async () => {
-          const nextImages = [...latestImagesRef.current, newImage];
-          const payload = {
-            notes,
-            images: nextImages,
-            custom_fields: normalizedCustomFields,
-          };
-          const saveFn = saveStateFn ?? ((p) => updateCurrentState(pieceId, p));
-          const result = await saveFn(payload);
-          const savedState = saveStateFn
-            ? (result.history.find((ps) => ps.id === initialPieceState.id) ??
-              result.current_state)
-            : result.current_state;
-          latestImagesRef.current = savedState.images;
-          dispatch({
-            type: "replace_base_state",
-            pieceState: savedState,
-          });
-          onSaved(result);
-        });
-      imageSaveQueueRef.current = queuedSave;
-      queuedSave
-        .catch(() => setImageError("Failed to save image. Please try again."))
-        .finally(() => {
-          if (imageSaveQueueRef.current === queuedSave) {
-            setSavingImage(false);
-          }
-        });
-    },
-    [
-      initialPieceState.id,
-      normalizedCustomFields,
-      notes,
-      onSaved,
-      pieceId,
-      saveStateFn,
-    ],
-  );
+  const {
+    uploadError,
+    widgetLoading,
+    savingImage,
+    imageError,
+    handleUploadWidgetClick,
+  } = useImageSaveQueue({
+    pieceId,
+    initialStateId: initialPieceState.id,
+    saveStateFn,
+    notes,
+    normalizedCustomFields,
+    images,
+    onSaved,
+    dispatch,
+    readOnly,
+  });
 
   useEffect(() => {
     pieceDetailSaveStatus?.publishWorkflowStatus({
@@ -352,100 +196,6 @@ export default function WorkflowState({
     autosave.status,
     pieceDetailSaveStatus,
   ]);
-
-  async function handleUploadWidgetClick() {
-    if (readOnly) {
-      return;
-    }
-    setUploadError(null);
-    setWidgetLoading(true);
-    // Keeps the iframe hidden until it is ready (removed on display-changed: shown).
-    const hideStyle = document.createElement("style");
-    hideStyle.textContent = 'iframe[title="Upload Widget"] { opacity: 0; }';
-    document.head.appendChild(hideStyle);
-    // On iOS PWA with viewport-fit=cover the status bar overlays the page.
-    // Push the widget iframe below it so its header controls are reachable.
-    // This style must persist for the widget's lifetime (not removed with hideStyle).
-    const safeAreaStyle = document.createElement("style");
-    safeAreaStyle.textContent =
-      'iframe[title="Upload Widget"] { top: env(safe-area-inset-top) !important; height: calc(100dvh - env(safe-area-inset-top)) !important; }';
-    document.head.appendChild(safeAreaStyle);
-
-    const cleanupWidgetStyles = () => {
-      hideStyle.remove();
-      safeAreaStyle.remove();
-    };
-    const uploadWidget = await openCloudinaryUploadWidget({
-      messages: {
-        configError: "Failed to load upload configuration. Please try again.",
-        unavailableError:
-          "Upload widget is not available in this browser. Please try again.",
-        signatureError: "Failed to sign upload. Please try again.",
-        uploadError: "Upload failed. Please try again.",
-      },
-      widgetOptions: {
-        sources: ["local", "camera"],
-        multiple: true,
-        resourceType: "image",
-        styles: {
-          palette: {
-            window: theme.palette.background.paper,
-            windowBorder: theme.palette.divider,
-            tabIcon: theme.palette.primary.main,
-            menuIcons: theme.palette.text.secondary,
-            textDark: theme.palette.text.primary,
-            textLight: theme.palette.text.secondary,
-            link: theme.palette.primary.main,
-            action: theme.palette.primary.dark,
-            inactiveTabIcon: theme.palette.text.disabled,
-            error: theme.palette.error.main,
-            inProgress: theme.palette.primary.main,
-            complete: theme.palette.success.main,
-            sourceBg: theme.palette.background.default,
-          },
-          frame: { background: "#00000000" },
-        } as { palette: Record<string, string> },
-      },
-      callbacks: {
-        onError: (message) => {
-          setWidgetLoading(false);
-          cleanupWidgetStyles();
-          setUploadError(message);
-        },
-        onDisplayChange: (state) => {
-          if (state === "shown") {
-            setWidgetLoading(false);
-            hideStyle.remove();
-            const iframe = document.querySelector(
-              'iframe[title="Upload Widget"]',
-            );
-            if (iframe instanceof HTMLElement) {
-              iframe.style.transition = "opacity 0.15s ease-in";
-              iframe.style.opacity = "1";
-            }
-          } else if (state === "hidden" || state === "destroyed") {
-            safeAreaStyle.remove();
-          }
-        },
-        onSuccess: (result, config) => {
-          const newImage = {
-            url: result.info.secure_url,
-            caption: "",
-            cloudinary_public_id: result.info.public_id,
-            cloud_name: config.cloud_name,
-            crop: null,
-            width: result.info.width ?? null,
-            height: result.info.height ?? null,
-          };
-          saveUploadedImage(newImage);
-        },
-      },
-    });
-    if (!uploadWidget) {
-      setWidgetLoading(false);
-      cleanupWidgetStyles();
-    }
-  }
 
   function handleFieldChange(name: string, value: string) {
     dispatch({ type: "set_custom_field", name, value });
