+++ web/src/stories/LightboxFooter.stories.tsx
@@ -1,0 +1,101 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { fn } from "@storybook/test";
+import LightboxFooter from "../components/LightboxFooter";
+import type { PiecePhotoGalleryImage } from "../components/PiecePhotoGallery";
+
+/**
+ * LightboxFooter renders the control bar at the bottom of the ImageLightbox.
+ *
+ * Rationale: Extracted from PiecePhotoGallery.tsx (Issue #406) to own caption
+ * editing and photo-move state independently of the gallery route logic.
+ *
+ * Edge cases:
+ * - Read-only: caption and move buttons are disabled.
+ * - Caption editing: inline TextField replaces the caption chip.
+ * - Move menu: lists other states the photo can be moved to.
+ * - No caption: shows an "Add caption" button placeholder.
+ * - Thumbnail: icon button highlighted when image is the current thumbnail.
+ */
+const meta = {
+  title: "Components/LightboxFooter",
+  component: LightboxFooter,
+  parameters: { layout: "centered", backgrounds: { default: "dark" } },
+  tags: ["autodocs"],
+} satisfies Meta<typeof LightboxFooter>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+const activeImage: PiecePhotoGalleryImage = {
+  url: "https://res.cloudinary.com/demo/image/upload/sample.jpg",
+  caption: "",
+  cloudinary_public_id: "sample",
+  cloud_name: "demo",
+  crop: null,
+  image_id: "img1",
+  stateLabel: "Trimmed",
+  stateId: "s1",
+  editableCurrentStateIndex: 0,
+};
+
+const editableImages = [
+  {
+    url: activeImage.url,
+    caption: "",
+    cloudinary_public_id: "sample",
+    cloud_name: "demo",
+    crop: null,
+    image_id: "img1",
+  },
+];
+
+const pieceStates = [
+  { id: "s1", label: "Trimmed" },
+  { id: "s2", label: "Bisque fired" },
+  { id: "s3", label: "Glaze fired" },
+];
+
+export const Default: Story = {
+  args: {
+    activeImage,
+    editableCurrentStateImages: editableImages,
+    editableCurrentStateIndex: 0,
+    pieceStates,
+    persistCurrentStateImages: fn(),
+    moveImageFn: fn(),
+    onPieceUpdated: fn(),
+    galleryPath: "/pieces/p1/photos",
+    outerState: {},
+    settingThumbnail: false,
+    isCurrentThumbnail: false,
+    cropAvailable: true,
+    onCrop: fn(),
+    onSetAsThumbnail: fn(),
+  },
+};
+
+export const WithCaption: Story = {
+  args: {
+    ...Default.args,
+    activeImage: { ...activeImage, caption: "Trimmed foot ring, 1.2lbs" },
+    editableCurrentStateImages: [
+      { ...editableImages[0], caption: "Trimmed foot ring, 1.2lbs" },
+    ],
+  },
+};
+
+export const CurrentThumbnail: Story = {
+  args: {
+    ...Default.args,
+    isCurrentThumbnail: true,
+  },
+};
+
+export const ReadOnly: Story = {
+  args: {
+    ...Default.args,
+    editableCurrentStateIndex: null,
+    moveImageFn: undefined,
+    onSetAsThumbnail: undefined,
+  },
+};
